Move ecr.IsRepositoryPolicyUpToDate to awsclients.IsPolicyUpToDate
This function is useful for comparing JSON policy documents for
multiple clients/controllers, so we move is to a shared package.
