+++ pkg/clients/ecr/repository_policy_test.go
@@ -20,10 +20,7 @@ var (
 	boolCheck   = true
 	testID      = "id"
 	policy      = `{"Statement":[{"Action":"ecr:ListImages","Effect":"Allow","Principal":"*"}],"Version":"2012-10-17"}`
-	cpxPolicy   = `{"Statement":[{"Action":"ecr:ListImages","Effect":"Allow","Principal":{"AWS":["arn:aws:iam::111122223333:userARN","111122223334","arn:aws:iam::111122223333:roleARN"]}}],"Version":"2012-10-17"}`
-	// Note: different sort order of principals than input above
-	cpxRemPolicy = `{"Statement":[{"Action":"ecr:ListImages","Effect":"Allow","Principal":{"AWS":["111122223334","arn:aws:iam::111122223333:userARN","arn:aws:iam::111122223333:roleARN"]}}],"Version":"2012-10-17"}`
-	params       = v1alpha1.RepositoryPolicyParameters{
+	params      = v1alpha1.RepositoryPolicyParameters{
 		Policy: &v1alpha1.RepositoryPolicyBody{
 			Version: "2012-10-17",
 			Statements: []v1alpha1.RepositoryPolicyStatement{
@@ -196,65 +193,6 @@ func TestLateInitializePolicy(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			LateInitializeRepositoryPolicy(tc.parameters, tc.policyOutput)
 			if diff := cmp.Diff(tc.want, tc.parameters); diff != "" {
-				t.Errorf("r: -want, +got:\n%s", diff)
-			}
-		})
-	}
-}
-
-func TestIsRepositoryPolicyUpToDate(t *testing.T) {
-	type args struct {
-		local  string
-		remote string
-	}
-
-	cases := map[string]struct {
-		args args
-		want bool
-	}{
-		"SameFields": {
-			args: args{
-				local:  "{\"testone\": \"one\", \"testtwo\": \"two\"}",
-				remote: "{\"testtwo\": \"two\", \"testone\": \"one\"}",
-			},
-			want: true,
-		},
-		"SameFieldsRealPolicy": {
-			args: args{
-				local:  policy,
-				remote: `{"Statement":[{"Effect":"Allow","Action":"ecr:ListImages","Principal":"*"}],"Version":"2012-10-17"}`,
-			},
-			want: true,
-		},
-		"DifferentFields": {
-			args: args{
-				local:  "{\"testone\": \"one\", \"testtwo\": \"two\"}",
-				remote: "{\"testthree\": \"three\", \"testone\": \"one\"}",
-			},
-			want: false,
-		},
-		"SameFieldsPrincipalPolicy": {
-			args: args{
-				local:  cpxPolicy,
-				remote: cpxRemPolicy,
-			},
-			want: true,
-		},
-		"SameFieldsNumericPrincipals": {
-			args: args{
-				// This is to test that our slice sorting does not
-				// panic with unexpected value types.
-				local:  `{"Statement":[{"Effect":"Allow","Action":"ecr:ListImages","Principal":[2,1,"foo","bar"]}],"Version":"2012-10-17"}`,
-				remote: `{"Statement":[{"Effect":"Allow","Action":"ecr:ListImages","Principal":[2,1,"bar","foo"]}],"Version":"2012-10-17"}`,
-			},
-			want: true,
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			got := IsRepositoryPolicyUpToDate(&tc.args.local, &tc.args.remote)
-			if diff := cmp.Diff(tc.want, got); diff != "" {
 				t.Errorf("r: -want, +got:\n%s", diff)
 			}
 		})
